+++ src/pages/Search/components/RecentItem/RecentItem.module.scss
