+++ src/pages/Search/components/RollingBanner/RollingBanner.module.scss
@@ -1,0 +1,76 @@
+@use "src/utils/styles/mediaQuery.scss" as media;
+
+$highlight: #ff7f23;
+$default: #c4c4c4;
+
+.banner {
+  position: relative;
+  overflow: hidden;
+  margin-top: 24px;
+  height: 32px;
+  width: 100%;
+  font-size: 16px;
+
+  @include media.media-breakpoint-down(mobile) {
+    margin-left: 10px;
+    width: 120%;
+  }
+
+  &::before,
+  &::after {
+    content: "";
+    position: absolute;
+    top: 0;
+    height: 100%;
+    width: 38px;
+    pointer-events: none;
+    z-index: 1;
+  }
+
+  &::before {
+    left: 0;
+    background: linear-gradient(to right, #ffffff, transparent);
+  }
+
+  &::after {
+    right: 0;
+    background: linear-gradient(to left, #ffffff, transparent);
+  }
+
+  &__tag-lists {
+    display: flex;
+    white-space: nowrap;
+    overflow: hidden;
+
+    @keyframes text-loop {
+      0% { transform: translate3d(0, 0, 0); }
+      100% { transform: translate3d(-100%, 0, 0); }
+    }
+
+    @include media.media-breakpoint-down(mobile) {
+      max-width: 80%;
+      font-size: 16px;
+    }
+  }
+
+  &__tag-list {
+    display: flex;
+    animation: text-loop 10s linear infinite;
+    list-style: none;
+  }
+
+  &__tag {
+    border-radius: 15px;
+    margin-right: 10px;
+    padding: 3px 7px;
+    color: $default;
+    border: 1px solid $default;
+
+    &:hover {
+      cursor: default;
+      color: $highlight;
+      border: 1px solid $highlight;
+      background-color: rgba(246 191 84 / 20%);
+    }
+  }
+}
