+++ src/pages/Search/components/SearchBar/SearchBar.module.scss
@@ -1,49 +1,22 @@
 @use "src/utils/styles/mediaQuery.scss" as media;
 
-$highlight: #ff7f23;
-$default: #c4c4c4;
-
 .search {
   width: 100%;
   height: 100vh;
-  font-size: 32px;
-
-  &__sentence {
-    word-break: keep-all;
-    height: 80px;
-    padding-top: 168px;
-    padding-left: 24px;
-    font-size: 40px;
-    font-weight: 300;
-    line-height: 49px;
-    white-space: pre-wrap;
-    color: #000000;
-
-    strong {
-      color: #ff7f23;
-    }
-
-    @include media.media-breakpoint-down(mobile) {
-      max-width: 767px;
-      padding-top: 64px;
-      margin: 0 auto;
-    }
-  }
 }
 
 .search-bar {
   position: relative;
   display: flex;
-  width: 98%;
-  border-radius: 100px;
-  border: none;
+  flex-grow: 1;
+  width: 100%;
   height: 62px;
   margin-top: 34px;
-  background: #ffffff;
+  border: none;
+  border-radius: 100px;
+  background-color: transparent;
   box-shadow: 2px 3px 12px 1px rgba(0 0 0 / 10%);
-  padding-left: 15px;
-  z-index: 3;
-  flex-grow: 1;
+  z-index: 2;
 
   @include media.media-breakpoint-down(mobile) {
     width: 95%;
@@ -65,10 +38,11 @@ $default: #c4c4c4;
     border: none;
     font-size: 24px;
     line-height: 15px;
-    padding-left: 24px;
+    margin-left: 24px;
 
     &::placeholder {
       color: #666666;
+      font-weight: 500;
       font-size: 24px;
 
       @include media.media-breakpoint-down(mobile) {
@@ -103,77 +77,5 @@ $default: #c4c4c4;
     border: none;
     background: none;
     cursor: pointer;
-  }
-}
-
-.banner {
-  position: relative;
-  overflow: hidden;
-  margin-top: 24px;
-  height: 32px;
-  width: 100%;
-  font-size: 16px;
-
-  @include media.media-breakpoint-down(mobile) {
-    margin-left: 10px;
-    width: 120%;
-  }
-
-  &::before,
-  &::after {
-    content: "";
-    position: absolute;
-    top: 0;
-    height: 100%;
-    width: 38px;
-    pointer-events: none;
-    z-index: 2;
-  }
-
-  &::before {
-    left: 0;
-    background: linear-gradient(to right, #ffffff, transparent);
-  }
-
-  &::after {
-    right: 0;
-    background: linear-gradient(to left, #ffffff, transparent);
-  }
-
-  &__tag-lists {
-    display: flex;
-    white-space: nowrap;
-    overflow: hidden;
-
-    @keyframes text-loop {
-      0% { transform: translate3d(0, 0, 0); }
-      100% { transform: translate3d(-100%, 0, 0); }
-    }
-
-    @include media.media-breakpoint-down(mobile) {
-      max-width: 80%;
-      font-size: 16px;
-    }
-  }
-
-  &__tag-list {
-    display: flex;
-    animation: text-loop 10s linear infinite;
-    list-style: none;
-  }
-
-  &__tag {
-    border-radius: 15px;
-    margin-right: 10px;
-    padding: 3px 7px;
-    color: $default;
-    border: 1px solid $default;
-
-    &:hover {
-      cursor: default;
-      color: $highlight;
-      border: 1px solid $highlight;
-      background-color: rgba(246 191 84 / 20%);
-    }
   }
 }
