+++ src/pages/Search/components/SearchBar/SearchInput.tsx
@@ -12,12 +12,20 @@ export default function SearchInput({
   text, onChange, onSubmit,
 }: Props) {
   return (
-    <label title="검색어 입력" aria-label="검색어 입력" className={styles['search-bar']} htmlFor="searchBarInput">
-      <form onSubmit={onSubmit} className={styles['search-bar__form']}>
+    <label
+      className={styles['search-bar']}
+      title="검색어 입력"
+      aria-label="검색어 입력"
+      htmlFor="search-bar-input"
+    >
+      <form
+        className={styles['search-bar__form']}
+        onSubmit={onSubmit}
+      >
         <input
           className={styles['search-bar__input']}
           placeholder="검색어를 입력해주세요."
-          id="searchBarInput"
+          id="search-bar-input"
           autoComplete="off"
           value={text}
           onChange={onChange}
