+++ src/pages/Search/components/Sentence/Sentence.tsx
@@ -1,6 +1,6 @@
 import { useRef } from 'react';
 
-import styles from './SearchBar.module.scss';
+import styles from './Sentence.module.scss';
 
 interface SentenceProps {
   subText: JSX.Element[];
@@ -11,7 +11,7 @@ export default function Sentence({ subText }: SentenceProps) {
   const sentence = subText[textIndex.current];
 
   return (
-    <h1 className={styles.search__sentence}>
+    <h1 className={styles.sentence}>
       {sentence}
     </h1>
   );
